feat: extract drm code into backend, interfaced through vtable
