+++ rendering/rendering.h
@@ -2,16 +2,12 @@
 
 #include "canvas.h"
 
-#include <stddef.h>
-#include <stdint.h>
+struct rendering_vtable {
+	void *(*rendering_init)(void);
+	void (*rendering_cleanup)(void *r_ctx);
+	void (*rendering_ctx_log)(const void *r_ctx);
+	void (*rendering_show)(void *r_ctx, struct canvas *);
+	struct canvas *(*canvas_init)(void *r_ctx);
+};
 
-struct rendering_ctx;
-
-struct rendering_ctx *rendering_init(void);
-void rendering_cleanup(struct rendering_ctx *);
-
-void rendering_ctx_log(const struct rendering_ctx *);
-
-struct canvas *canvas_init(struct rendering_ctx *);
-
-void rendering_show(struct rendering_ctx *, struct canvas *);
+struct rendering_vtable drm_backend_new(void);
